+++ TagCloud/ColorPicker/IColorPicker.cs
@@ -4,6 +4,6 @@ namespace TagCloud.ColorPicker
 {
     public interface IColorPicker
     {
-        Color AdjustColor(Color baseColor, int usingFrequency);
+        Color AdjustColor(Color baseColor, float usingFrequency);
     }
 }
